fix Dockerfile
The docker build failed, because the build tools are not available in `node:alpine`.
The updated script installs the build tools, runs npm install and then removes the build tools to free space.

Also, the CMD is updated to `npm start` instead of `node index.js`. This way, the start command is managed by the `package.json`.